refactor: Refactor test helpers to allow writing tests that do not use NGitLab.Mock
Separate FakeContext into multiple classes for different purposes.

- FakeContext not longer implements IGitLabContext and can be used when only a test double for a ICakeContext is required
- NGitLabMockContext implements IGitLabContext and will return a provider that uses a mocked GitLab client created via NGitLab.Mock
