+++ test/Cake.GitLab.Test/TestHelpers/NGitLabMockContext.cs
@@ -1,0 +1,18 @@
+﻿using Cake.GitLab.Testing;
+using Xunit;
+
+namespace Cake.GitLab.Test;
+
+public class NGitLabMockContext : FakeContext, IGitLabCakeContext
+{
+    public NGitLabMockGitLabProvider GitLab { get; }
+
+    /// <inheritdoc />
+    IGitLabProvider IGitLabCakeContext.GitLab => GitLab;
+
+
+    public NGitLabMockContext(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
+    {
+        GitLab = new NGitLabMockGitLabProvider(this);
+    }
+}
